backend/requirements.txt
# === コアフレームワーク ===
Django==5.2.6
wagtail==7.0.3

# === API・連携 ===
djangorestframework==3.16.1
django-cors-headers==4.4.0
drf-spectacular==0.27.2

# === データ管理 ===
django-import-export==4.2.0
tablib[html,ods,xls,xlsx,yaml]==3.7.0

# === 機能強化 ===
django-filter==24.3
django-taggit==5.0.1
django-redis==5.4.0
django-autoslug==1.9.9
django-simple-history==3.10.1

# === Wagtail拡張 ===
wagtail-modeladmin==2.2.0
wagtail-cache==3.0.0
wagtail-metadata==5.0.0

# === インフラ ===
psycopg2-binary==2.9.10
redis==5.1.0
pillow==11.0.0
gunicorn==22.0.0

# === 開発ツール ===
django-debug-toolbar==4.4.6
django-extensions==3.2.3
ipython==8.26.0
